Change input for 'jumlah' to parseFloat

+++ public/js/master-produk/add-modal.js
@@ -196,7 +196,7 @@ $(function () {
         const idx = $(this).data("idx");
         const harga =
             parseInt($(`input[name="harga_bahan[]"]`).eq(idx).val()) || 0;
-        const jumlah = parseInt($(this).val()) || 0;
+        const jumlah = parseFloat($(this).val()) || 0;
 
         // Pastikan array terdefinisi
         if (
